Initial screen layout refactoring
Adds in the necessary screens to match the UI requirements and removes older proof-of-concept screens.
