+++ lib/routes.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:frequencypay/pages/forgot_password_page.dart';
+import 'package:frequencypay/pages/login/forgot_password_page.dart';
 import 'package:frequencypay/pages/home_page.dart';
-import 'package:frequencypay/pages/login_page.dart';
-import 'package:frequencypay/pages/registration_page.dart';
-import 'package:frequencypay/pages/welcome_page.dart';
+import 'package:frequencypay/pages/login/login_page.dart';
+import 'package:frequencypay/pages/login/registration_page.dart';
+import 'package:frequencypay/pages/login/welcome_page.dart';
 
 final routes = {
   '/home': (BuildContext context) => new HomePage(),
